# Use Line Filters in StringLabelFilter