--- a/pkg/logql/log/label_filter_test.go
+++ b/pkg/logql/log/label_filter_test.go
@@ -3,6 +3,7 @@ package log
 import (
 	"reflect"
 	"sort"
+	"strings"
 	"testing"
 	"time"
 
@@ -357,7 +358,7 @@ func TestStringLabelFilter(t *testing.T) {
 
 	tests := []struct {
 		name        string
-		filter      *StringLabelFilter
+		filter      LabelFilterer
 		labels      labels.Labels
 		shouldMatch bool
 	}{
@@ -417,6 +418,67 @@ func TestStringLabelFilter(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			_, ok := tc.filter.Process(0, []byte("sample log line"), NewBaseLabelsBuilder().ForLabels(tc.labels, tc.labels.Hash()))
 			assert.Equal(t, tc.shouldMatch, ok)
+		})
+	}
+}
+
+var result bool
+
+func BenchmarkLineLabelFilters(b *testing.B) {
+	line := []byte("line")
+	fixture := strings.Join([]string{
+		"foo", "foobar", "bar", "foobuzz", "buzz", "f", "  ", "fba", "foofoofoo", "b", "foob", "bfoo", "FoO",
+		"foo, 世界", allunicode(), "fooÏbar",
+	}, ",")
+	lbl := NewBaseLabelsBuilder().ForLabels(labels.Labels{
+		{"foo", fixture},
+	}, 0)
+
+	for _, test := range []struct {
+		re string
+	}{
+		// regex we intend to support.
+		{"foo"},
+		{"(foo)"},
+		{"(foo|ba)"},
+		{"(foo|ba|ar)"},
+		{"(foo|(ba|ar))"},
+		{"foo.*"},
+		{".*foo.*"},
+		{"(.*)(foo).*"},
+		{"(foo.*|.*ba)"},
+		{"(foo.*|.*bar.*)"},
+		{".*foo.*|bar"},
+		{".*foo|bar"},
+		{"(?:.*foo.*|bar)"},
+		{"(?P<foo>.*foo.*|bar)"},
+		{".*foo.*|bar|buzz"},
+		{".*foo.*|bar|uzz"},
+		{"foo|bar|b|buzz|zz"},
+		{"f|foo|foobar"},
+		{"f.*|foobar.*|.*buzz"},
+		{"((f.*)|foobar.*)|.*buzz"},
+		{".*"},
+		{".*|.*"},
+		{".*||||"},
+		{""},
+		{"(?i)foo"},
+		{"(?i)界"},
+		{"(?i)ïB"},
+		{"(?:)foo|fatal|exception"},
+		{"(?i)foo|fatal|exception"},
+		{"(?i)f|foo|foobar"},
+		{"(?i)f|fatal|e.*"},
+		{"(?i).*foo.*"},
+	} {
+		b.Run(test.re, func(b *testing.B) {
+			matcher := labels.MustNewMatcher(labels.MatchRegexp, "foo", test.re)
+			f := NewStringLabelFilter(matcher)
+			b.ResetTimer()
+
+			for i := 0; i < b.N; i++ {
+				_, result = f.Process(0, line, lbl)
+			}
 		})
 	}
 }