We've done a ton of work to optimize regexes in line filters but nothing in label filters. To avoid duplicating the work or reworking all of our filter types, this PR creates a new label filter that has an optimized line filter. It uses that against the label value to determine if they match.

Note on Benchmarks: the input string for this is 1114195 bytes long. The `not` test tells us that our approach is quite a lot slower when the literal pattern isn't present in the input. The size of the input really emphasizes it. 

Benchmarks:
```
name                                        old time/op    new time/op        delta
LineLabelFilters/foo-8                        50.6ns ± 9%        14.3ns ± 1%        -71.67%  (p=0.000 n=10+8)
LineLabelFilters/not-8                        15.7ns ± 1%     64165.6ns ± 2%    +407766.77%  (p=0.000 n=10+10)
LineLabelFilters/(foo)-8                      58.4ns ± 1%        14.8ns ±10%        -74.68%  (p=0.000 n=9+9)
LineLabelFilters/(foo|ba)-8                   64.9ns ± 0%        18.2ns ± 2%        -71.90%  (p=0.000 n=7+9)
LineLabelFilters/(foo|ba|ar)-8                74.8ns ± 7%        21.4ns ± 2%        -71.42%  (p=0.000 n=10+8)
LineLabelFilters/(foo|(ba|ar))-8              65.1ns ± 1%        18.1ns ± 2%        -72.22%  (p=0.000 n=8+8)
LineLabelFilters/foo.*-8                       882ns ± 4%          15ns ±11%        -98.30%  (p=0.000 n=9+10)
LineLabelFilters/.*foo-8                      11.5ns ± 8%        14.9ns ± 7%        +30.39%  (p=0.000 n=10+10)
LineLabelFilters/.*foo.*-8                    5.46µs ± 7%        0.01µs ± 3%        -99.74%  (p=0.000 n=10+9)
LineLabelFilters/(.*)(foo).*-8                6.26µs ± 7%        0.01µs ± 2%        -99.77%  (p=0.000 n=10+9)
LineLabelFilters/(foo.*|.*ba)-8               5.58µs ± 3%        0.02µs ± 8%        -99.67%  (p=0.000 n=9+10)
LineLabelFilters/(foo.*|.*bar.*)-8            7.02µs ± 7%        0.02µs ± 8%        -99.73%  (p=0.000 n=10+10)
LineLabelFilters/.*foo.*|bar-8                5.53µs ± 6%        0.02µs ± 2%        -99.67%  (p=0.000 n=10+8)
LineLabelFilters/.*foo|bar-8                  3.73µs ± 6%        0.02µs ± 3%        -99.52%  (p=0.000 n=10+9)
LineLabelFilters/(?:.*foo.*|bar)-8            5.43µs ± 8%        0.02µs ± 8%        -99.65%  (p=0.000 n=10+10)
LineLabelFilters/(?P<foo>.*foo.*|bar)-8       5.71µs ± 3%        0.02µs ±10%        -99.68%  (p=0.000 n=9+9)
LineLabelFilters/.*foo.*|bar|buzz-8           5.34µs ± 2%        0.02µs ± 3%        -99.65%  (p=0.000 n=10+9)
LineLabelFilters/.*foo.*|bar|uzz-8            5.36µs ± 1%        0.02µs ± 3%        -99.60%  (p=0.000 n=8+8)
LineLabelFilters/foo|bar|b|buzz|zz-8          68.8ns ± 8%        22.6ns ± 7%        -67.19%  (p=0.000 n=10+10)
LineLabelFilters/f|foo|foobar-8               77.1ns ± 1%        18.6ns ± 9%        -75.90%  (p=0.000 n=8+10)
LineLabelFilters/f.*|foobar.*|.*buzz-8        5.64µs ± 7%        0.02µs ±10%        -99.60%  (p=0.000 n=10+10)
LineLabelFilters/((f.*)|foobar.*)|.*buzz-8    6.29µs ± 8%        0.02µs ±10%        -99.65%  (p=0.000 n=10+10)
LineLabelFilters/.*-8                          963ns ± 8%           2ns ±14%        -99.77%  (p=0.000 n=10+10)
LineLabelFilters/.*|.*-8                      4.46µs ± 8%        0.00µs ± 8%        -99.95%  (p=0.000 n=10+10)
LineLabelFilters/.*||||-8                     2.99µs ± 1%        0.00µs ±16%        -99.93%  (p=0.000 n=8+10)
LineLabelFilters/#00-8                        43.5ns ± 8%         2.1ns ± 1%        -95.25%  (p=0.000 n=10+8)
LineLabelFilters/(?i)foo-8                    71.7ns ± 6%        13.1ns ± 1%        -81.73%  (p=0.000 n=10+8)
LineLabelFilters/(?i)界-8                      16.5ns ± 8%       150.2ns ± 8%       +809.55%  (p=0.000 n=10+10)
LineLabelFilters/(?i)ïB-8                     23.8ns ± 1%  14748979.8ns ± 7%  +62074730.81%  (p=0.000 n=8+10)
LineLabelFilters/(?:)foo|fatal|exception-8     185ns ± 2%          21ns ± 2%        -88.38%  (p=0.000 n=9+8)
LineLabelFilters/(?i)foo|fatal|exception-8    78.4ns ± 3%        20.2ns ± 2%        -74.23%  (p=0.000 n=9+9)
LineLabelFilters/(?i)f|foo|foobar-8           94.3ns ±16%        15.4ns ± 9%        -83.66%  (p=0.000 n=10+10)
LineLabelFilters/(?i)f|fatal|e.*-8            65.7ns ± 5%        18.2ns ± 3%        -72.24%  (p=0.000 n=9+8)
LineLabelFilters/(?i).*foo.*-8                6.04µs ± 8%        0.01µs ± 2%        -99.78%  (p=0.000 n=10+10)

name                                        old alloc/op   new alloc/op       delta
LineLabelFilters/foo-8                         0.00B              0.00B                ~     (all equal)
LineLabelFilters/not-8                         0.00B              0.00B                ~     (all equal)
LineLabelFilters/(foo)-8                       0.00B              0.00B                ~     (all equal)
LineLabelFilters/(foo|ba)-8                    0.00B              0.00B                ~     (all equal)
LineLabelFilters/(foo|ba|ar)-8                 0.00B              0.00B                ~     (all equal)
LineLabelFilters/(foo|(ba|ar))-8               0.00B              0.00B                ~     (all equal)
LineLabelFilters/foo.*-8                       0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*foo-8                       0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*foo.*-8                     0.00B              0.00B                ~     (all equal)
LineLabelFilters/(.*)(foo).*-8                 0.00B              0.00B                ~     (all equal)
LineLabelFilters/(foo.*|.*ba)-8                0.00B              0.00B                ~     (all equal)
LineLabelFilters/(foo.*|.*bar.*)-8             0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*foo.*|bar-8                 0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*foo|bar-8                   0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?:.*foo.*|bar)-8             0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?P<foo>.*foo.*|bar)-8        0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*foo.*|bar|buzz-8            0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*foo.*|bar|uzz-8             0.00B              0.00B                ~     (all equal)
LineLabelFilters/foo|bar|b|buzz|zz-8           0.00B              0.00B                ~     (all equal)
LineLabelFilters/f|foo|foobar-8                0.00B              0.00B                ~     (all equal)
LineLabelFilters/f.*|foobar.*|.*buzz-8         0.00B              0.00B                ~     (all equal)
LineLabelFilters/((f.*)|foobar.*)|.*buzz-8     0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*-8                          0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*|.*-8                       0.00B              0.00B                ~     (all equal)
LineLabelFilters/.*||||-8                      0.00B              0.00B                ~     (all equal)
LineLabelFilters/#00-8                         0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?i)foo-8                     0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?i)界-8                       0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?i)ïB-8                      0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?:)foo|fatal|exception-8     0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?i)foo|fatal|exception-8     0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?i)f|foo|foobar-8            0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?i)f|fatal|e.*-8             0.00B              0.00B                ~     (all equal)
LineLabelFilters/(?i).*foo.*-8                 0.00B              0.00B                ~     (all equal)

name                                        old allocs/op  new allocs/op      delta
LineLabelFilters/foo-8                          0.00               0.00                ~     (all equal)
LineLabelFilters/not-8                          0.00               0.00                ~     (all equal)
LineLabelFilters/(foo)-8                        0.00               0.00                ~     (all equal)
LineLabelFilters/(foo|ba)-8                     0.00               0.00                ~     (all equal)
LineLabelFilters/(foo|ba|ar)-8                  0.00               0.00                ~     (all equal)
LineLabelFilters/(foo|(ba|ar))-8                0.00               0.00                ~     (all equal)
LineLabelFilters/foo.*-8                        0.00               0.00                ~     (all equal)
LineLabelFilters/.*foo-8                        0.00               0.00                ~     (all equal)
LineLabelFilters/.*foo.*-8                      0.00               0.00                ~     (all equal)
LineLabelFilters/(.*)(foo).*-8                  0.00               0.00                ~     (all equal)
LineLabelFilters/(foo.*|.*ba)-8                 0.00               0.00                ~     (all equal)
LineLabelFilters/(foo.*|.*bar.*)-8              0.00               0.00                ~     (all equal)
LineLabelFilters/.*foo.*|bar-8                  0.00               0.00                ~     (all equal)
LineLabelFilters/.*foo|bar-8                    0.00               0.00                ~     (all equal)
LineLabelFilters/(?:.*foo.*|bar)-8              0.00               0.00                ~     (all equal)
LineLabelFilters/(?P<foo>.*foo.*|bar)-8         0.00               0.00                ~     (all equal)
LineLabelFilters/.*foo.*|bar|buzz-8             0.00               0.00                ~     (all equal)
LineLabelFilters/.*foo.*|bar|uzz-8              0.00               0.00                ~     (all equal)
LineLabelFilters/foo|bar|b|buzz|zz-8            0.00               0.00                ~     (all equal)
LineLabelFilters/f|foo|foobar-8                 0.00               0.00                ~     (all equal)
LineLabelFilters/f.*|foobar.*|.*buzz-8          0.00               0.00                ~     (all equal)
LineLabelFilters/((f.*)|foobar.*)|.*buzz-8      0.00               0.00                ~     (all equal)
LineLabelFilters/.*-8                           0.00               0.00                ~     (all equal)
LineLabelFilters/.*|.*-8                        0.00               0.00                ~     (all equal)
LineLabelFilters/.*||||-8                       0.00               0.00                ~     (all equal)
LineLabelFilters/#00-8                          0.00               0.00                ~     (all equal)
LineLabelFilters/(?i)foo-8                      0.00               0.00                ~     (all equal)
LineLabelFilters/(?i)界-8                        0.00               0.00                ~     (all equal)
LineLabelFilters/(?i)ïB-8                       0.00               0.00                ~     (all equal)
LineLabelFilters/(?:)foo|fatal|exception-8      0.00               0.00                ~     (all equal)
LineLabelFilters/(?i)foo|fatal|exception-8      0.00               0.00                ~     (all equal)
LineLabelFilters/(?i)f|foo|foobar-8             0.00               0.00                ~     (all equal)
LineLabelFilters/(?i)f|fatal|e.*-8              0.00               0.00                ~     (all equal)
LineLabelFilters/(?i).*foo.*-8                  0.00               0.00                ~     (all equal)
```