@jeschkies 2023-03-01
Nice find! Would you have a flame graph for the Unicode case? This is quite a drop.