@MasslessParticle 2023-03-01
@jeschkies Turns out it's super expensive to call `toLower` on unicode:

![image](https://user-images.githubusercontent.com/1413241/222259679-89542f3f-19a1-4425-abef-841f5dbc589a.png)