@MasslessParticle 2023-03-01
@karsten I dug into this today and discovered that Prometheus Label Semantics are different than line filters. Label matches are anchored to the beginning and ending of lines so there is no concept of "contains a literal" for labels. This PR originally broke those semantics so some operations that were essentially no-ops now took a ton of time.

I've fixed the regex simplification to account for the label semantics and the benchmarks make way more sense:

```
name                                        old time/op  new time/op  delta
LineLabelFilters/foo-8                      52.9ns ± 9%  48.5ns ± 5%   -8.43%  (p=0.000 n=10+10)
LineLabelFilters/(foo)-8                    59.2ns ± 9%  51.1ns ±12%  -13.64%  (p=0.000 n=10+10)
LineLabelFilters/(foo|ba)-8                 63.6ns ± 2%  65.9ns ±10%     ~     (p=0.400 n=9+10)
LineLabelFilters/(foo|ba|ar)-8              71.7ns ± 5%  68.8ns ± 4%   -4.13%  (p=0.001 n=10+9)
LineLabelFilters/(foo|(ba|ar))-8            65.1ns ± 4%  70.1ns ± 9%   +7.66%  (p=0.000 n=9+10)
LineLabelFilters/foo.*-8                     900ns ± 7%    15ns ±11%  -98.33%  (p=0.000 n=10+10)
LineLabelFilters/.*foo.*-8                  5.15µs ± 3%  0.01µs ± 4%  -99.71%  (p=0.000 n=9+9)
LineLabelFilters/(.*)(foo).*-8              5.96µs ± 4%  0.02µs ± 8%  -99.74%  (p=0.000 n=10+10)
LineLabelFilters/(foo.*|.*ba)-8             5.69µs ±10%  0.02µs ± 8%  -99.67%  (p=0.000 n=10+10)
LineLabelFilters/(foo.*|.*bar.*)-8          6.60µs ± 2%  0.02µs ± 9%  -99.71%  (p=0.000 n=8+10)
LineLabelFilters/.*foo.*|bar-8              5.39µs ±13%  5.32µs ± 4%     ~     (p=0.529 n=10+8)
LineLabelFilters/.*foo|bar-8                3.65µs ± 9%  3.63µs ± 2%     ~     (p=0.234 n=10+9)
LineLabelFilters/(?:.*foo.*|bar)-8          5.37µs ± 8%  5.61µs ± 7%   +4.49%  (p=0.020 n=10+10)
LineLabelFilters/(?P<foo>.*foo.*|bar)-8     5.77µs ± 8%  5.51µs ± 7%     ~     (p=0.105 n=10+10)
LineLabelFilters/.*foo.*|bar|buzz-8         5.41µs ± 8%  0.02µs ±10%  -99.65%  (p=0.000 n=10+10)
LineLabelFilters/.*foo.*|bar|uzz-8          5.40µs ± 8%  5.43µs ± 2%     ~     (p=0.203 n=10+8)
LineLabelFilters/foo|bar|b|buzz|zz-8        66.8ns ± 2%  72.9ns ± 7%   +9.09%  (p=0.000 n=8+10)
LineLabelFilters/f|foo|foobar-8             78.3ns ± 2%  19.1ns ± 9%  -75.58%  (p=0.000 n=9+10)
LineLabelFilters/f.*|foobar.*|.*buzz-8      5.50µs ± 7%  0.02µs ±11%  -99.59%  (p=0.000 n=10+10)
LineLabelFilters/((f.*)|foobar.*)|.*buzz-8  6.18µs ± 7%  0.02µs ±10%  -99.64%  (p=0.000 n=10+10)
LineLabelFilters/.*-8                        930ns ± 8%     2ns ±16%  -99.76%  (p=0.000 n=9+10)
LineLabelFilters/.*|.*-8                    4.48µs ± 7%  0.00µs ± 2%  -99.95%  (p=0.000 n=10+8)
LineLabelFilters/.*||||-8                   2.99µs ± 3%  0.00µs ±15%  -99.93%  (p=0.000 n=10+10)
LineLabelFilters/#00-8                      48.0ns ± 9%   2.2ns ± 8%  -95.50%  (p=0.000 n=10+10)
LineLabelFilters/(?i)foo-8                  74.3ns ±16%  66.4ns ±10%  -10.68%  (p=0.006 n=10+10)
LineLabelFilters/(?i)界-8                    17.4ns ±13%  16.1ns ±22%     ~     (p=0.063 n=10+10)
LineLabelFilters/(?i)ïB-8                   24.2ns ±10%  23.6ns ± 4%     ~     (p=0.136 n=9+9)
LineLabelFilters/(?:)foo|fatal|exception-8   186ns ± 8%    71ns ±10%  -62.09%  (p=0.000 n=10+10)
LineLabelFilters/(?i)foo|fatal|exception-8  81.1ns ± 8%  82.4ns ± 8%     ~     (p=0.218 n=10+10)
LineLabelFilters/(?i)f|foo|foobar-8         88.3ns ± 3%  15.7ns ±13%  -82.22%  (p=0.000 n=8+10)
LineLabelFilters/(?i)f|fatal|e.*-8          66.8ns ± 8%  18.2ns ± 2%  -72.73%  (p=0.000 n=10+8)
LineLabelFilters/(?i).*foo.*-8              5.75µs ± 2%  0.01µs ± 8%  -99.76%  (p=0.000 n=8+10)

```